Refactor TraderDesk into modular package
