+++ traderdesk/__init__.py
@@ -1,0 +1,5 @@
+"""Core package for the TraderDesk application."""
+
+from .app import main
+
+__all__ = ["main"]
