+++ traderdesk/constants.py
@@ -1,0 +1,4 @@
+"""Application wide constants."""
+
+COMMISSION = 0.0001
+SLIPPAGE = 0.0001
