+++ traderdesk/ui/__init__.py
@@ -1,0 +1,5 @@
+"""Qt user interface components for TraderDesk."""
+
+from .main_window import TraderDesk
+
+__all__ = ["TraderDesk"]
